requirements.txt
fastapi
uvicorn[standard]
alpaca-trade-api
transformers
torch
newsapi-python
python-dotenv
